+++ src/main/java/racingcar/race/Car.java
@@ -1,8 +1,9 @@
-package racingcar.util;
+package racingcar.race;
 
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import racingcar.util.PickRandomNumber;
 
 public class Car {
